packages/archetypeAnnotater/requirements.txt
numpy
umap-learn
hdbscan
